+++ .storybook/config.js
@@ -12,7 +12,9 @@ function loadStories() {
 
 setOptions({
   name: 'Oyez Editor',
+  sortStoriesByKind: true,
   url: '#',
+  showDownPanel: true
 });
 
 configure(loadStories, module);