+++ src/index.js
@@ -1,5 +1,6 @@
 export * from './components/editor'
 export * from './components/editor-block'
+export * from './components/editor-quill'
 export * from './components/example-menu'
 export * from './components/example-image'
 export * from './components/example-input'