+++ CHANGES.md
@@ -11,3 +11,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Initial Project Setup with Lint, Test and Bundle Setup
 - Basic Editor Components Setup (Editor, EditorBlock and Example Components)
 - Storybook implemented to showcase Editor Components
+- Quill Editor Component implemented (Basic Setup)